Renaming class RegisterIo to FpgaIo
