+++ src/ArduinoViperFpga.cpp
@@ -21,7 +21,7 @@ extern void enableFpgaClock(); /* Defined within MKRVIDOR4000/variant.cpp */
  **************************************************************************************/
 
 ArduinoViperFpga::ArduinoViperFpga(ViperFpga::SpiSelectFunc select, ViperFpga::SpiDeselectFunc deselect, ViperFpga::SpiTransferFunc transfer)
-: _io_reg(select, deselect, transfer)
+: _io(select, deselect, transfer)
 {
 
 }
@@ -38,23 +38,23 @@ ArduinoViperFpga::Status ArduinoViperFpga::begin()
 
 uint8_t ArduinoViperFpga::getRevNum()
 {
-  return _io_reg.read(ViperFpga::Register::FPGA_REV_NUM);
+  return _io.read(ViperFpga::Register::FPGA_REV_NUM);
 }
 
 ArduinoViperFpga::Status ArduinoViperFpga::setIntensityRed(uint8_t const red)
 {
-  _io_reg.write(ViperFpga::Register::RGB_LED_RED_INTENSITY, red);
+  _io.write(ViperFpga::Register::RGB_LED_RED_INTENSITY, red);
   return Status::OK;
 }
 
 ArduinoViperFpga::Status ArduinoViperFpga::setIntensityGreen(uint8_t const green)
 {
-  _io_reg.write(ViperFpga::Register::RGB_LED_GREEN_INTENSITY, green);
+  _io.write(ViperFpga::Register::RGB_LED_GREEN_INTENSITY, green);
   return Status::OK;
 }
 
 ArduinoViperFpga::Status ArduinoViperFpga::setIntensityBlue(uint8_t const blue)
 {
-  _io_reg.write(ViperFpga::Register::RGB_LED_BLUE_INTENSITY, blue);
+  _io.write(ViperFpga::Register::RGB_LED_BLUE_INTENSITY, blue);
   return Status::OK;
 }
